Remove zoning module
