+++ lib/parcel_web/controllers/land_uses_controller.ex
@@ -1,11 +1,11 @@
 defmodule ParcelWeb.LandUsesController do
   use ParcelWeb, :controller
-  alias Parcel.Zoning
+  alias Parcel.Zoning.ZoningDistrictLandUseCondition
 
   action_fallback(ParcelWeb.FallbackController)
 
   def index(conn, %{"address" => address}) do
-    json(conn, Zoning.land_use_summary(address))
+    json(conn, ZoningDistrictLandUseCondition.land_use_summary(address))
   end
 
   def index(_conn, _params), do: {:error, :bad_request, "Missing address parameter."}
